requirements.txt
flask
groq
selenium
requests
python-dotenv